Setting correctly datasource to production version

+++ backend/src/data-source.ts
@@ -1,28 +1,45 @@
 import { DataSource } from "typeorm"
 import "reflect-metadata";
 import dotenv from 'dotenv';
+import fs = require('fs');
 
 dotenv.config();
 
-let dataSource:DataSource;
+let dataSource = new DataSource({
+    type: "postgres",
+    host: process.env.DB_HOST,
+    port: Number(process.env.DB_PORT),
+    username: process.env.DB_USERNAME,
+    password: process.env.DB_PASSWORD,
+    database: process.env.DB_DATABASE,
+    synchronize: true,
+    logging: true,
+    entities: [__dirname + '/contexts/**/infrastructure/persistence/entities/*.{ts,js}'],
+    subscribers: [],
+    migrations: [__dirname + '/db/migrations/*.{ts,js}'],
+    dropSchema: false
+});
 
 switch (process.env.NODE_ENV) {
     case 'prod':
-        dataSource = new DataSource({
-            type: "postgres",
-            host: process.env.DB_HOST,
-            port: Number(process.env.DB_PORT),
-            username: process.env.DB_USERNAME,
-            password: process.env.DB_PASSWORD,
-            database: process.env.DB_DATABASE,
-            synchronize: true,
-            logging: true,
-            entities: [__dirname + '/contexts/**/infrastructure/persistence/entities/*.{ts,js}'],
-            subscribers: [],
-            migrations: [__dirname + '/db/migrations/*.{ts,js}'],
-            dropSchema: false,
-            ssl: { ca: process.env.SSL_CERT }
-        });
+        const pathSSLCert = process.env.SSL_CERT;
+        if (pathSSLCert) {
+            dataSource = new DataSource({
+                type: "postgres",
+                host: process.env.DB_HOST,
+                port: Number(process.env.DB_PORT),
+                username: process.env.DB_USERNAME,
+                password: process.env.DB_PASSWORD,
+                database: process.env.DB_DATABASE,
+                synchronize: true,
+                logging: true,
+                entities: [__dirname + '/contexts/**/infrastructure/persistence/entities/*.{ts,js}'],
+                subscribers: [],
+                migrations: [__dirname + '/db/migrations/*.{ts,js}'],
+                dropSchema: false,
+                ssl: {ca: fs.readFileSync(pathSSLCert, 'utf-8').toString() }
+            });
+        }
         break;
 
     case 'test':
@@ -37,23 +54,6 @@ switch (process.env.NODE_ENV) {
             logging: false,
             entities: [__dirname + '/contexts/**/infrastructure/persistence/entities/*.{ts,js}'],
             subscribers: []
-        });
-        break;
-
-    default:
-        dataSource = new DataSource({
-            type: "postgres",
-            host: process.env.DB_HOST,
-            port: Number(process.env.DB_PORT),
-            username: process.env.DB_USERNAME,
-            password: process.env.DB_PASSWORD,
-            database: process.env.DB_DATABASE,
-            synchronize: true,
-            logging: true,
-            entities: [__dirname + '/contexts/**/infrastructure/persistence/entities/*.{ts,js}'],
-            subscribers: [],
-            migrations: [__dirname + '/db/migrations/*.{ts,js}'],
-            dropSchema: false
         });
         break;
 }
